+++ app/(site)/[eventSlug]/proposals/voting-buttons.tsx
@@ -129,6 +129,7 @@ export function VotingButtons({
       <HoverTooltip
         text={votingEnabled ? "Interested" : votingDisabledText}
         visible={true}
+        unavailable={!votingEnabled}
       >
         <button
           type="button"
@@ -136,7 +137,6 @@ export function VotingButtons({
             ${large ? "w-16 h-16 sm:w-20 sm:h-20 flex flex-col items-center justify-center" : "px-1 py-1"}
             ${votingEnabled ? "" : "opacity-50 cursor-not-allowed grayscale"}
             ${votes.some((vote) => vote.proposalId === proposalId && vote.choice === VoteChoice.interested && vote.guestId === currentUserId) ? "bg-blue-200" : "bg-white"}`}
-          disabled={!votingEnabled}
           onClick={(e) => handleVote(VoteChoice.interested, e)}
         >
           <div className={large ? "text-sm sm:text-lg mb-1" : ""}>❤️</div>
@@ -146,6 +146,7 @@ export function VotingButtons({
       <HoverTooltip
         text={votingEnabled ? "Maybe" : votingDisabledText}
         visible={true}
+        unavailable={!votingEnabled}
       >
         <button
           type="button"
@@ -153,7 +154,6 @@ export function VotingButtons({
             ${large ? "w-16 h-16 sm:w-20 sm:h-20 flex flex-col items-center justify-center" : "px-1 py-1"}
             ${votingEnabled ? "" : "opacity-50 cursor-not-allowed grayscale"}
             ${votes.some((vote) => vote.proposalId === proposalId && vote.choice === VoteChoice.maybe && vote.guestId === currentUserId) ? "bg-blue-200" : "bg-white"}`}
-          disabled={!votingEnabled}
           onClick={(e) => handleVote(VoteChoice.maybe, e)}
         >
           <div className={large ? "text-sm sm:text-lg mb-1" : ""}>⭐</div>
@@ -163,6 +163,7 @@ export function VotingButtons({
       <HoverTooltip
         text={votingEnabled ? "Skip" : votingDisabledText}
         visible={true}
+        unavailable={!votingEnabled}
       >
         <button
           type="button"
@@ -170,7 +171,6 @@ export function VotingButtons({
             ${large ? "w-16 h-16 sm:w-20 sm:h-20 flex flex-col items-center justify-center" : "px-1 py-1"}
             ${votingEnabled ? "" : "opacity-50 cursor-not-allowed grayscale"}
             ${votes.some((vote) => vote.proposalId === proposalId && vote.choice === VoteChoice.skip && vote.guestId === currentUserId) ? "bg-blue-200" : "bg-white"}`}
-          disabled={!votingEnabled}
           onClick={(e) => handleVote(VoteChoice.skip, e)}
         >
           <div className={large ? "text-sm sm:text-lg mb-1" : ""}>👋🏽</div>